feat: remove Toggle button and move TooltipTrigger to /components (#692), :recycle:
